requirements.txt
PyOpenGL
PyOpenGL_accelerate
glfw
numpy
trimesh
imgui[glfw]
Pillow
pyglm